requirements.txt
rembg>=2.0.50
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
onnxruntime>=1.16.0